// src/app/flows/fines/fines-acc/fines-acc-minor-creditor-add-amend-convert/interfaces/fines-acc-minor-creditor-add-amend-convert-state.interface.ts
export interface IFinesAccMinorCreditorAddAmendConvertState {
  facc_minor_creditor_creditor_type: string | null;
  facc_minor_creditor_title: string | null;
  facc_minor_creditor_forenames: string | null;
  facc_minor_creditor_surname: string | null;
  facc_minor_creditor_company_name: string | null;
  facc_minor_creditor_address_line_1: string | null;
  facc_minor_creditor_address_line_2: string | null;
  facc_minor_creditor_address_line_3: string | null;
  facc_minor_creditor_post_code: string | null;
  facc_minor_creditor_pay_by_bacs: boolean | null;
  facc_minor_creditor_bank_account_name: string | null;
  facc_minor_creditor_bank_sort_code: string | null;
  facc_minor_creditor_bank_account_number: string | null;
  facc_minor_creditor_bank_account_reference: string | null;
}